+++ src/main/java/wooteco/subway/service/LineService.java
@@ -3,10 +3,10 @@ package wooteco.subway.service;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
-import wooteco.subway.dao.Entity.SectionEntity;
-import wooteco.subway.dao.LineRepository;
-import wooteco.subway.dao.SectionRepository;
-import wooteco.subway.dao.StationRepository;
+import wooteco.subway.reopository.Entity.SectionEntity;
+import wooteco.subway.reopository.LineRepository;
+import wooteco.subway.reopository.SectionRepository;
+import wooteco.subway.reopository.StationRepository;
 import wooteco.subway.domain.Line;
 import wooteco.subway.domain.Section;
 import wooteco.subway.domain.Sections;
