+++ src/test/java/wooteco/subway/reopository/SectionDaoTest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dao;
+package wooteco.subway.reopository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -8,10 +8,11 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
-import wooteco.subway.dao.Entity.SectionEntity;
+import wooteco.subway.reopository.Entity.SectionEntity;
 import wooteco.subway.domain.Line;
 import wooteco.subway.domain.Section;
 import wooteco.subway.domain.Station;
+import wooteco.subway.reopository.dao.SectionDao;
 
 @JdbcTest
 @Import({SectionRepository.class, SectionDao.class})
