+++ src/test/java/wooteco/subway/service/LineServiceTest.java
@@ -10,10 +10,10 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
-import wooteco.subway.dao.Entity.SectionEntity;
-import wooteco.subway.dao.LineRepository;
-import wooteco.subway.dao.SectionRepository;
-import wooteco.subway.dao.StationRepository;
+import wooteco.subway.reopository.Entity.SectionEntity;
+import wooteco.subway.reopository.LineRepository;
+import wooteco.subway.reopository.SectionRepository;
+import wooteco.subway.reopository.StationRepository;
 import wooteco.subway.domain.Line;
 import wooteco.subway.domain.Section;
 import wooteco.subway.domain.Station;
